requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
strawberry-graphql==0.187.0
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.7
kafka-python==2.0.2
prometheus-client==0.17.1
pydantic==1.10.12
python-dotenv==1.0.0
confluent-kafka==2.12.2
typing-extensions==4.8.0
anyio==4.11.0
starlette>=0.27.0,<0.28.0